pyproject.toml
[project]
name = "neurodriver"
version = "0.1.0"
description = "End-to-End Autonomous Driving with IL + RL in CARLA"
requires-python = ">=3.10"

dependencies = [
    # Core ML
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "torchaudio>=2.1.0",
    
    # RL
    "stable-baselines3>=2.1.0",
    "gymnasium>=0.29.0",
    
    # Data & Config
    "numpy>=1.24.0",
    "h5py>=3.9.0",
    "Pillow>=10.0.0",
    "opencv-python-headless>=4.8.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    
    # Logging & Viz
    "wandb>=0.15.0",
    "matplotlib>=3.7.0",
    "tqdm>=4.65.0",
    
    # Export
    "onnx>=1.14.0",
    "onnxruntime>=1.16.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "ipykernel>=6.25.0",
    "jupyter>=1.0.0",
]
carla = [
    "carla==0.9.15",
]
diffusion = [
    "diffusers>=0.21.0",
    "transformers>=4.33.0",
    "accelerate>=0.23.0",
]

[tool.setuptools.packages.find]
include = ["neurodriver*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
